pyproject.toml
[tool.poetry]
name = "electronics-shop-project"
version = "0.1.0"
description = ""
authors = ["kirillskyeng <kirill.netreba@skyeng.ru>"]
readme = "README.md"
packages = [{include = "electronics_shop_project"}]

[tool.poetry.dependencies]
python = "^3.9.1"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
